[ANDROSDK-1319] Create SettingAppService
